Hopefully the finishing touch
Functionality
=============
» When calculating the amount needed to scroll, we now parse the partHeight as a float instead of an integer. This is to take into account the fact different browsers have slightly different heights based on how they are rendered, for example in firefox the height of certain parts are '48.4px' where in chrome they are '48px' and in Edge they are '48.16px'. This should just make it slightly more accurate.

+++ src/ui.js
@@ -176,7 +176,7 @@ const UI = {
           } else {
             const prevPart = i - 1;
             for(x = 0; x <= prevPart; x++) {
-              const partHeight = parseInt(window.getComputedStyle(partList[x], null).getPropertyValue('height'));
+              const partHeight = parseFloat(window.getComputedStyle(partList[x], null).getPropertyValue('height'));
               scrollAmt += partHeight;
             }
           }
